+++ src/ast_variable_ref.cpp
@@ -8,7 +8,7 @@ void VariableReference::EmitRISC(std::ostream &stream, Context &context) const
     std::string name = ss.str();
     Variable *v = context.FindVariable(name);
     stream << "# Load variable \"" << v->name << "\"\n";
-    stream << "lw a0," << (long)v->offset << "(s0)\n";
+    stream << "lw a0," << v->offset << "(s0)\n";
 }
 
 void VariableReference::Print(std::ostream &stream) const
